# ascon: generate 128a code