The perf improvements are mostly from getting rid of needless
spills inside the permutation routines. Inlining P6 inside
encryptBlocks128aGeneric, decryptBlocks128aGeneric, and
additionalData128aGeneric accounts for about 10 MB/s.

```
name           old time/op   new time/op   delta
Seal1K_128a-8   3.37µs ± 0%   2.22µs ± 0%  -34.00%  (p=0.000 n=10+8)
Open1K_128a-8   3.25µs ± 0%   2.23µs ± 0%  -31.48%  (p=0.000 n=9+10)
Seal8K_128a-8   25.6µs ± 0%   16.8µs ± 0%  -34.16%  (p=0.000 n=9+9)
Open8K_128a-8   24.6µs ± 0%   16.8µs ± 0%  -31.66%  (p=0.000 n=9+10)
Seal1K_128-8    4.93µs ± 0%   3.48µs ± 0%  -29.37%  (p=0.000 n=9+8)
Open1K_128-8    4.80µs ± 0%   3.49µs ± 0%  -27.25%  (p=0.000 n=6+10)
Seal8K_128-8    37.9µs ± 0%   27.5µs ± 0%  -27.38%  (p=0.000 n=9+8)
Open8K_128-8    36.8µs ± 0%   26.9µs ± 0%  -27.05%  (p=0.000 n=8+9)

name           old speed     new speed     delta
Seal1K_128a-8  304MB/s ± 0%  461MB/s ± 0%  +51.50%  (p=0.000 n=10+9)
Open1K_128a-8  315MB/s ± 0%  459MB/s ± 0%  +45.92%  (p=0.000 n=9+9)
Seal8K_128a-8  320MB/s ± 0%  487MB/s ± 0%  +51.87%  (p=0.000 n=9+9)
Open8K_128a-8  333MB/s ± 0%  487MB/s ± 0%  +46.33%  (p=0.000 n=9+10)
Seal1K_128-8   208MB/s ± 0%  294MB/s ± 0%  +41.59%  (p=0.000 n=9+8)
Open1K_128-8   213MB/s ± 0%  293MB/s ± 0%  +37.46%  (p=0.000 n=7+10)
Seal8K_128-8   216MB/s ± 0%  298MB/s ± 0%  +37.70%  (p=0.000 n=9+8)
Open8K_128-8   222MB/s ± 0%  305MB/s ± 0%  +37.08%  (p=0.000 n=8+9)
```

Signed-off-by: Eric Lagergren <eric@ericlagergren.com>